Cache the audio and video
On Linux, the cached copies are stored in $XDG_CACHE_HOME (or its
default, ~/.cache) according to the XDG specification.
On Mac, the cache dir is ~/Library/Caches.
On Windows, the cache dir remains /tmp.
In all cases, a new subfolder under the cache dir named "astley" is made
and the files are stored there.

The integrity of the cached copies is verified using MD5 hashes.

Mac has md5 and Linux and Cygwin should have md5sum as part of
coreutils. This catches the case where the file was only partially
downloaded, and if so, it will refetch the file.